Implement bar widget for tuner reference frequency, show detected frequency in widget_tuner

+++ app/inc/glcd_widget.h
@@ -110,8 +110,8 @@ typedef struct PEAKMETER_T {
 typedef struct TUNER_T {
     float frequency;
     const char *note;
-    int8_t cents;
-    float ref_freq;
+    int16_t cents;
+    int16_t ref_freq;
     uint8_t input;
 } tuner_t;
